backend/ai/main.py
from fastapi import FastAPI
from fastapi.middleware.cors import CORSMiddleware
import uvicorn
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Create FastAPI app
app = FastAPI(
    title="StratIQ AI Service",
    description="AI Trading Strategy Analysis and Learning Engine",
    version="1.0.0"
)

# Add CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],  # Configure for production
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

@app.get("/")
async def root():
    return {
        "message": "StratIQ AI Service is running!",
        "version": "1.0.0",
        "status": "healthy"
    }

@app.get("/health")
async def health_check():
    return {
        "status": "OK",
        "service": "StratIQ AI",
        "endpoints": [
            "/strategy/analyze",
            "/recall/insights", 
            "/risk/calculate",
            "/market/sentiment"
        ]
    }

@app.post("/strategy/analyze")
async def analyze_strategy(strategy_data: dict):
    # TODO: Implement strategy analysis
    return {
        "message": "Strategy analysis coming soon",
        "received_data": strategy_data
    }

@app.get("/recall/insights")
async def get_recall_insights():
    # TODO: Implement Recall leaderboard analysis
    return {
        "message": "Recall insights coming soon",
        "top_performers": [],
        "patterns": []
    }

@app.post("/risk/calculate")
async def calculate_risk(portfolio_data: dict):
    # TODO: Implement risk calculation
    return {
        "message": "Risk calculation coming soon",
        "risk_score": 0.0,
        "recommendations": []
    }

@app.get("/market/sentiment")
async def get_market_sentiment():
    # TODO: Implement sentiment analysis
    return {
        "message": "Market sentiment analysis coming soon",
        "sentiment": "neutral",
        "confidence": 0.0
    }

if __name__ == "__main__":
    port = int(os.getenv("PORT", 8000))
    uvicorn.run(
        "main:app", 
        host="0.0.0.0", 
        port=port, 
        reload=True,
        log_level="info"
    )