backend/ai/requirements.txt
pandas
numpy
scikit-learn
fastapi
uvicorn
requests
python-dotenv
openai
